requirements.txt
fastapi
uvicorn
pymupdf
pdfminer.six
sentence-transformers
faiss-cpu
transformers
torch
langchain
networkx
pandas
numpy
scikit-learn
streamlit
python-dotenv
tqdm